GraphClient: move query Options code to own file
